requirements.txt
boto3
configparser
argparse
python-dateutil
